ppo_cart_pole/pyproject.toml
[project]
name = "ppo-cart-pole"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium[other]>=1.2.3",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "hydra-core>=1.3",
    "omegaconf",
    "matplotlib>=3.10.8",
    "tqdm>=4.67.3",
    "pygame>=2.6.1",
]
